# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Daycore v2 API
  version: "2.4.0"
  description: >
    Daycore v2 backend (Go). Autonomous AI day-planning + mood companion.
    Anonymous-first: POST /api/session/init issues an httpOnly, signed session
    cookie (dc_sid) that authorizes all data endpoints. Optional email/OAuth
    sign-in sets an httpOnly JWT cookie (dc_auth); logout revokes all issued JWTs
    (server-side token version). Same-origin web frontends MUST send cookies
    (fetch: credentials:"include"). Native apps and separately-deployed frontends
    use header credentials instead (dual-rail, cookies stay supported): the
    signed session token in X-Session-Token (returned by /api/session/init with
    {"tokenInBody":true}) and/or the JWT in Authorization: Bearer (returned by
    login/register with {"tokenInBody":true}). Check GET /api/version (apiVersion
    contract major) before use — matching rules in FRONTEND_HANDOFF.md.
    The browser extension authenticates to /api/import/* with X-Import-Token;
    admin endpoints use X-Admin-Token.
    Errors are {error, message}. The single SSE endpoint is /api/ai/companion —
    see FRONTEND_HANDOFF.md §B for the v2 frame protocol (NOT the legacy [DONE]).
servers:
  - url: http://localhost:8080
    description: Local (Go serves the API; nginx proxies /api in production)
...
security:
  - sessionCookie: []
  - sessionToken: []
  - bearerAuth: []
paths:
...
  /api/v2/ai/companion:
    post:
      operationId: companion
      tags: [ai]
      summary: Companion agent over SSE v2 (tool-calls frame protocol).
      description: >
        Returns text/event-stream. Each frame is `data: {json}\n\n` discriminated
        by `type`: delta | reasoning | tool_start | tool_result | decision_card |
        error | done. Comment heartbeats `: ping`. NO `[DONE]` sentinel. See
        FRONTEND_HANDOFF.md §B. With threadId the server loads AND persists thread
        history; client roles are whitelisted to user/assistant.
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { message: { type: string, description: "May be empty when attachmentIds is not — dropping in a photo with no words is a complete thing to say." }, timezone: { type: string }, location: { type: string, maxLength: 128, description: "Where the device is, free text. Fills in or updates a device-detected location; never overwrites one the user set on the settings page. Used only by the briefs — the chat tool takes a location as an argument." }, assistantName: { type: string }, threadId: { type: string }, attachmentIds: { type: array, maxItems: 20, items: { type: string }, description: "Ids from POST /api/files. Must be this session's unsent uploads; they become part of the persisted user message. Files the chosen model cannot read are described to it by name rather than dropped." }, conversationHistory: { type: array, items: { type: object, properties: { role: { type: string, enum: [user, assistant] }, content: { type: string } } } } } } } } }
      responses: { '200': { description: SSE stream (text/event-stream), content: { text/event-stream: { schema: { type: string }, example: 'data: {"type":"delta","text":"好的"}\n\ndata: {"type":"done"}\n\n' } } } }
...
  /api/v2/ai/auto-plan:
    post:
      operationId: autoPlan
      tags: [ai]
      summary: Autonomous multi-day planning from imported materials.
      requestBody: { content: { application/json: { schema: { type: object, properties: { from: { type: string }, to: { type: string }, mode: { type: string, enum: [keep_manual, replace_all] }, extra: { type: string } } } } } }
      responses: { '200': { description: OK }, '400': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: dc_sid }
    authCookie: { type: apiKey, in: cookie, name: dc_auth }
    sessionToken: { type: apiKey, in: header, name: X-Session-Token, description: "Same signed value as the dc_sid cookie, for clients without a cookie jar. Get it from POST /api/session/init {tokenInBody:true}." }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: "Same JWT as the dc_auth cookie. Get it from login/register {tokenInBody:true}; logout revokes all issued JWTs (token-version bump)." }
    importToken: { type: apiKey, in: header, name: X-Import-Token, description: "Session-scoped token for the browser extension's direct push." }
    adminToken: { type: apiKey, in: header, name: X-Admin-Token, description: "Admin endpoints." }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: "Stable code the frontend translates (e.g. no_session, bad_request, rate_limited, token_expired)." }
        message: { type: string, description: "Localized fallback message." }
...